feature: Config for max number of attached photos and max pixel size
Key: homerep.service-request.images.max-num-photos
Default: 10
Configures the max number of photos which can be attached to a single
service request.
Attempting to POST /api/customer/service_request/{id}/attach with more
than this amount results in an error
  {"timestamp": 123, "type": "max_pictures"}

Key: homerep.service-request.images.max-size-pixels
Default: 1920
Configures the max number of pixels in width and height that can be
attached to a service request. Pictures larger than this will be resized
to fit.