feature: Add rate limiting by IP to register and password_reset/send
RestController endpoints can be annotated with @RateLimited to limit
the number of times a client can use an endpoint. When the limit is
reached, an error of 429 Too Many Requests will be sent.
The rate-limit only affects valid requests.

Key: homerep.rate-limit.enabled
Default: true
Set to false to disable all rate-limiting.