feature: Check for conflicting Appointments properly
Unique constraints did not account for CANCELLED appointments and did
not allow for multiple service providers to create appointments for a
customer on a single day.
MySQL and H2 seem to have a CHECK constraint, but don't support the
sub-queries required to enforce this constraint.
So we have to check for conflicting appointment when creating and
confirming appointments.
Marked all enum fields in entities with Enumerated(EnumType.STRING) to
save the name of the enum value instead of its ordinal in the db.

For an UNCONFIRMED appointment get the CONFIRMED appointments that
conflict with the appointment's time period
GET /api/customer/appointments/{id}/conflicting
  OK [AppointmentModel, ...]

POST /api/customer/appointments/{id}/confirm
  No request body
  OK No response body
  400 {type: "non_existent_post"}
  400 {type: "conflicting_appointment",
       conflictingAppointments: [AppointmentModel, ...]}

POST /api/provider/service_requests/{service_request_id}/appointments
      /create
  {date: "2023-11-19", message: "..." or null}
  OK No response body
  400 {type: "non_existent_post"}
  400 {type: "missing_accepted_email_request"}
  400 {type: "conflicting_appointment",
       conflictingAppointments: [AppointmentModel, ...]}