fix: Placeholder data and profile info changes without clicking Update
Typing in the input fields for editing the account info would change the
information on the card and in the placeholder text even without
clicking the Update button.

Format phone number on profile card and placeholder.
Add email line to customer profile.
Redirect to login if the user is not logged in.
Redirect to the correct profile page if the user type does not match.
Use correct navbar in CustomerProfile